Move db module mocking to its own mock file
