+++ src/__tests__/ilmonneet.test.js
@@ -5,33 +5,6 @@ import gsheetHandler from "../gsheet/handler";
 import type { APIGatewayEvent } from "../common/types";
 import { getIlmosFromDatabase } from "../common/db/ilmos";
 
-jest.mock("../common/db", () => {
-  const data = {
-    ILMOS_TABLE: { Items: {} },
-    USERS_TABLE: { Items: {} }
-  };
-  return {
-    deleteItem: ({ TableName, Key }) => {
-      delete data[TableName].Items[Key];
-    },
-    put: ({ TableName, Item }) => {
-      const items = data[TableName].Items;
-      let key;
-      if (TableName === "ILMOS_TABLE") {
-        key = Item.date;
-      } else if (TableName === "USERS_TABLE") {
-        key = Item.id;
-      } else {
-        throw new Error(`Unknown table name "${TableName}"`);
-      }
-      items[key] = Item;
-    },
-    scan: ({ TableName }) => {
-      return { Items: Object.values(data[TableName].Items) };
-    }
-  };
-});
-
 test("gsheet POST -> /ilmonneet", async () => {
   await lambdaCall(gsheetHandler, [
     ["", "", "", "Soprano", "", "Alto", "", "Tenor", "Bass", ""],
